+++ src/renderer/components/message/MessageList.tsx
@@ -11,10 +11,7 @@ import { C, T } from '@deltachat/jsonrpc-client'
 import moment from 'moment'
 
 import { MessageWrapper } from './MessageWrapper'
-import ChatStore, {
-  useChatStore,
-  ChatStoreStateWithChatSet,
-} from '../../stores/chat'
+import ChatStore, { ChatStoreStateWithChatSet } from '../../stores/chat'
 import { getLogger } from '../../../shared/logger'
 import { KeybindAction } from '../../keybindings'
 import { selectedAccountId } from '../../ScreenController'
@@ -25,6 +22,7 @@ import { MessagesDisplayContext } from '../../contexts/MessagesDisplayContext'
 import useTranslationFunction from '../../hooks/useTranslationFunction'
 import useKeyBindingAction from '../../hooks/useKeyBindingAction'
 import { useReactionsBar } from '../ReactionsBar'
+import EmptyChatMessage from './EmptyChatMessage'
 
 const log = getLogger('render/components/message/MessageList')
 
@@ -407,7 +405,11 @@ export default function MessageList({
 
   return (
     <MessagesDisplayContext.Provider
-      value={{ context: 'chat_messagelist', chatId: chatStore.chat.id }}
+      value={{
+        context: 'chat_messagelist',
+        chatId: chatStore.chat.id,
+        isDeviceChat: chatStore.chat.isDeviceChat,
+      }}
     >
       <MessageListInner
         onScroll={onScroll}
@@ -528,7 +530,9 @@ export const MessageListInner = React.memo(
         onScroll={onScroll}
       >
         <ul>
-          {messageListItems.length === 0 && <EmptyChatMessage />}
+          {messageListItems.length === 0 && (
+            <EmptyChatMessage chatStore={chatStore} />
+          )}
           {activeView.map(messageId => {
             if (messageId.kind === 'dayMarker') {
               return (
@@ -646,40 +650,6 @@ function JumpDownButton({
         </div>
       </div>
     </>
-  )
-}
-
-function EmptyChatMessage() {
-  const tx = useTranslationFunction()
-  const chatStore = useChatStore()
-  const chat = chatStore.chat
-
-  if (!chat) {
-    throw new Error('no chat selected')
-  }
-
-  let emptyChatMessage = tx('chat_new_one_to_one_hint', [chat.name, chat.name])
-
-  if (chat.chatType === C.DC_CHAT_TYPE_BROADCAST) {
-    emptyChatMessage = tx('chat_new_broadcast_hint')
-  } else if (chat.chatType === C.DC_CHAT_TYPE_GROUP && !chat.isContactRequest) {
-    emptyChatMessage = chat.isUnpromoted
-      ? tx('chat_new_group_hint')
-      : tx('chat_no_messages')
-  } else if (chat.isSelfTalk) {
-    emptyChatMessage = tx('saved_messages_explain')
-  } else if (chat.isDeviceChat) {
-    emptyChatMessage = tx('device_talk_explain')
-  } else if (chat.isContactRequest) {
-    emptyChatMessage = tx('chat_no_messages')
-  }
-
-  return (
-    <li>
-      <div className='info-message big'>
-        <div className='bubble'>{emptyChatMessage}</div>
-      </div>
-    </li>
   )
 }