+++ src/renderer/contexts/MessagesDisplayContext.tsx
@@ -1,7 +1,7 @@
 import { createContext } from 'react'
 
 type MessagesDisplayContextValue =
-  | { context: 'chat_messagelist'; chatId: number }
+  | { context: 'chat_messagelist'; chatId: number; isDeviceChat: boolean }
   | { context: 'chat_map'; chatId: number }
   | {
       context: 'contact_profile_status'